Add comments, rename entities to more logical names, reorganize code
