+++ client/CMakeLists.txt
@@ -88,6 +88,9 @@ set(vcmiclientcommon_SRCS
 	media/CSoundHandler.cpp
 	media/CVideoHandler.cpp
 
+	netlag/NetworkLagCompensator.cpp
+	netlag/PackRollbackGeneratorVisitor.cpp
+
 	render/AssetGenerator.cpp
 	render/CAnimation.cpp
 	render/CBitmapHandler.cpp
@@ -184,7 +187,6 @@ set(vcmiclientcommon_SRCS
 
 	xBRZ/xbrz.cpp
 
-	AntilagServer.cpp
 	ArtifactsUIController.cpp
 	GameEngine.cpp
 	GameInstance.cpp
@@ -301,6 +303,12 @@ set(vcmiclientcommon_HEADERS
 	media/ISoundPlayer.h
 	media/IVideoPlayer.h
 
+	netlag/NetworkLagCompensator.h
+	netlag/NetworkLagPredictedPack.h
+	netlag/NetworkLagPredictionTestVisitor.h
+	netlag/NetworkLagReplyPrediction.h
+	netlag/PackRollbackGeneratorVisitor.h
+
 	render/AssetGenerator.h
 	render/CAnimation.h
 	render/CBitmapHandler.h
@@ -408,7 +416,6 @@ set(vcmiclientcommon_HEADERS
 	xBRZ/xbrz.h
 	xBRZ/xbrz_tools.h
 
-	AntilagServer.h
 	ArtifactsUIController.h
 	CMT.h
 	CPlayerInterface.h
