+++ client/ClientCommandManager.h
@@ -45,6 +45,9 @@ class ClientCommandManager //take mantis #2292 issue about account if thinking a
 	// Redraw the current screen
 	void handleRedrawCommand();
 
+	// Enable or disable network lag compensation
+	void handleAntilagCommand(std::istringstream& singleWordBuffer);
+
 	// Extracts all translateable game texts into Translation directory, separating files on per-mod basis
 	void handleTranslateGameCommand(bool onlyMissing);
